+++ include/sway/desktop/launcher.h
@@ -1,0 +1,14 @@
+#ifndef _SWAY_LAUNCHER_H
+#define _SWAY_LAUNCHER_H
+
+#include <stdlib.h>
+
+struct sway_workspace *root_workspace_for_pid(pid_t pid);
+
+void root_record_workspace_pid(pid_t pid);
+
+void root_remove_workspace_pid(pid_t pid);
+
+void root_rename_pid_workspaces(const char *old_name, const char *new_name);
+
+#endif
