+++ sway/commands/exec_always.c
@@ -8,6 +8,7 @@
 #include "sway/commands.h"
 #include "sway/config.h"
 #include "sway/server.h"
+#include "sway/desktop/launcher.h"
 #include "sway/tree/container.h"
 #include "sway/tree/root.h"
 #include "sway/tree/workspace.h"
