+++ sway/commands/rename.c
@@ -7,6 +7,7 @@
 #include "sway/config.h"
 #include "sway/ipc-server.h"
 #include "sway/output.h"
+#include "sway/desktop/launcher.h"
 #include "sway/tree/container.h"
 #include "sway/tree/workspace.h"
 #include "sway/tree/root.h"
