+++ sway/meson.build
@@ -21,6 +21,7 @@ sway_sources = files(
 	'desktop/surface.c',
 	'desktop/transaction.c',
 	'desktop/xdg_shell.c',
+	'desktop/launcher.c',
 
 	'input/input-manager.c',
 	'input/cursor.c',
